#18 Fix: Fix Data JPA Repository Mapping
- Data Jpa에서 잘못된 제네릭 매핑을 수정했습니다.

+++ linknamu/src/main/java/com/kakao/linknamu/bookmark/BookmarkJPARepository.java
@@ -3,6 +3,6 @@ package com.kakao.linknamu.bookmark;
 import com.kakao.linknamu.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookmarkJPARepository extends JpaRepository<User, Integer> {
+public interface BookmarkJPARepository extends JpaRepository<User, Long> {
 
 }
