+++ linknamu/src/main/java/com/kakao/linknamu/user/repository/UserJPARepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserJPARepository extends JpaRepository<User, Integer> {
+public interface UserJPARepository extends JpaRepository<User, Long> {
 
     Optional<User> findByEmail(String email);
 }
